requirements.txt
pandas>=2.2.3
empyrical>=0.5.5
numpy>=1.26.4
optuna>=4.2.0
pytz>=2024.2
python-dateutil>=2.9.0.post0
scikit-learn>=1.6.1
sportsball>=0.2.8
feature-engine>=1.8.2
tqdm>=4.67.1
openskill>=6.0.2
pandarallel>=1.6.5
torch>=2.5.1
catboost>=1.2.7
scipy>=1.15.1
joblib>=1.4.2
venn-abers>=1.4.6
matplotlib>=3.10.0
pyfolio-reloaded>=0.9.8
fullmonte>=0.0.4
skfolio>=0.7.0
